requirements/test.txt
-r base.txt # includes the base.txt requirements file

Faker==0.8.4
nose==1.3.7
django-nose==1.4.5
coverage==4.4.1